Fix another pointer->int cast

+++ linuxdoom-1.10/r_data.c
@@ -634,7 +634,7 @@ void R_InitColormaps (void)
     lump = W_GetNumForName("COLORMAP"); 
     length = W_LumpLength (lump) + 255; 
     colormaps = Z_Malloc (length, PU_STATIC, 0); 
-    colormaps = (byte *)( ((int)colormaps + 255)&~0xff); 
+    colormaps = (byte *)( ((size_t)colormaps + 255)&~0xff); 
     W_ReadLump (lump,colormaps); 
 }
